refactor: rename actionCard file to action
